refactor(ts): extract timeline/layouts.ts from store
Move node-rect / effect-layout computeds, buildVisibleEndMap, and the
coordinate-space converters (toTimelineSpace/toViewportSpace/getNodeRect)
into a useTimelineLayouts(deps) composable, mirroring the optimizer. Wired
at end-of-setup so its compiledTimeline dep is already declared. Adds the
TimelineRect domain type. No behavior change.
